refactor(app): split useRunningStepCounts
