+++ app/src/resources/protocols/hooks/useRunningStepCounts.ts
@@ -3,6 +3,7 @@ import { useMostRecentCompletedAnalysis } from '/app/resources/runs'
 import { useLastRunProtocolCommand } from './useLastRunProtocolCommand'
 
 import type { CommandsData } from '@opentrons/api-client'
+import { getRunningStepCountsFrom } from '/app/resources/protocols'
 
 export interface StepCounts {
   /* Excludes "fixit" commands. Returns null if the step is not found. */
@@ -35,21 +36,5 @@ export function useRunningStepCounts(
     commandsData ?? null
   )
 
-  const lastRunAnalysisCommandIndex = analysisCommands.findIndex(
-    c => c.key === lastRunCommandNoFixit?.key
-  )
-
-  const currentStepNumberByAnalysis =
-    lastRunAnalysisCommandIndex === -1 ? null : lastRunAnalysisCommandIndex + 1
-
-  const hasRunDiverged =
-    lastRunCommandNoFixit?.key == null || currentStepNumberByAnalysis == null
-
-  const totalStepCount = !hasRunDiverged ? analysisCommands.length : null
-
-  return {
-    currentStepNumber: currentStepNumberByAnalysis,
-    totalStepCount,
-    hasRunDiverged,
-  }
+  return getRunningStepCountsFrom(analysisCommands, lastRunCommandNoFixit)
 }
